requirements.txt
tensorflow_gpu==1.5.0
spacy==2.0.11
allennlp==0.4.2
numpy==1.14.2
